+++ Util/fft/FFTW_Transform.h
@@ -18,21 +18,21 @@ namespace FTransform {
   public:
 
     //! Constructor
-    FFTW_Plan (unsigned nfft, const std::string& call);
+    FFTW_Plan (size_t nfft, const std::string& call);
 
     //! Destructor
     ~FFTW_Plan ();
 
-    static int fcc1d (unsigned nfft, float* dest, const float* src);
-    static int bcc1d (unsigned nfft, float* dest, const float* src);
-    static int frc1d (unsigned nfft, float* dest, const float* src);
-    static int bcr1d (unsigned nfft, float* dest, const float* src);
+    static int fcc1d (size_t nfft, float* dest, const float* src);
+    static int bcc1d (size_t nfft, float* dest, const float* src);
+    static int frc1d (size_t nfft, float* dest, const float* src);
+    static int bcr1d (size_t nfft, float* dest, const float* src);
 
     //! Agent class
     class Agent : public PlanAgent<FFTW_Plan> {
     public:
       Agent () : PlanAgent<FFTW_Plan> ("FFTW", nfft) { }
-      FFTW_Plan* new_plan (unsigned nfft, const std::string& call);
+      FFTW_Plan* new_plan (size_t nfft, const std::string& call);
     };
 
   protected:
@@ -46,19 +46,19 @@ namespace FTransform {
 
   public:
     FFTW_Plan2 ();
-    FFTW_Plan2 (unsigned nx, unsigned ny, const std::string& call);
+    FFTW_Plan2 (size_t nx, size_t ny, const std::string& call);
     ~FFTW_Plan2 ();
-    void init (unsigned nx, unsigned ny, const std::string& call);
+    void init (size_t nx, size_t ny, const std::string& call);
 
     static void
-    fcc2d (unsigned nx, unsigned ny, float* dest, const float* src);
+    fcc2d (size_t nx, size_t ny, float* dest, const float* src);
     static void
-    bcc2d (unsigned nx, unsigned ny, float* dest, const float* src);
+    bcc2d (size_t nx, size_t ny, float* dest, const float* src);
 
     class Agent : public PlanAgent2<FFTW_Plan2> {
     public:
       Agent () : PlanAgent2<FFTW_Plan2> ("FFTW", nfft) { }
-      FFTW_Plan2* new_plan (unsigned nx, unsigned ny, const std::string& call);
+      FFTW_Plan2* new_plan (size_t nx, size_t ny, const std::string& call);
     };
 
   protected:
